# packages/openapi/src/openapi.yaml
openapi: 3.0.3
info:
  title: PeridotID API
  version: 0.1.0
  description: |
    Gaming Identity Platform — Authentication, Identity, and Profile for the Peridot ecosystem.

    ## Base URL
    All endpoints are prefixed with `/v1`.

    | Environment | Base URL |
    |---|---|
    | Production | `https://api.pid.peridotvault.com/v1` |
    | Local dev | `http://localhost:3301/v1` |

    ## Authentication
    Authentication is **cookie-based**. After a successful Google sign-in the API sets two
    `HttpOnly` cookies:

    - `pid_access` — short-lived JWT access token (15 min), sent on every request.
    - `pid_refresh` — long-lived rotating refresh token (30 days), only sent to `/v1/auth`.

    Browsers handle these automatically; non-browser clients must send
    `Cookie: pid_access=<token>` on each request and call `/v1/auth/refresh` before the
    access token expires.
  contact:
    name: PeridotID
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.pid.peridotvault.com/v1
    description: Production
  - url: http://localhost:3301/v1
    description: Local development
...
paths:
...
  /apps:
    post:
      tags: [apps]
      summary: Register a third-party app
      description: |
        Registers an app for "Sign in with PeridotID" and returns its public
        `client_id`. Pass just a name — allowed origins are managed after via
        PATCH (the exact return URL is passed in code at login time).
        Authenticated — the caller becomes the app owner.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 60
                allowedOrigins:
                  type: array
                  maxItems: 20
                  description: Bare http(s) origins (no paths). Optional — managed after registration too.
                  items:
                    type: string
            example:
              name: "My Game"
      responses:
        "201":
          description: The registered app including its client_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PidApp"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [apps]
      summary: List my registered apps
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Apps owned by the caller
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PidApp"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /apps/{id}/secret:
    post:
      tags: [apps]
      summary: Generate or rotate an app's backend secret
      description: |
        Returns a `pidsk_…` secret **once** — only its SHA-256 hash is stored, so an
        existing secret can never be read back, only replaced. Afterwards, code
        exchange for this app requires the secret. Backend-only: never ship it in
        frontend code. Owners only.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The new secret (shown once) plus its identifying prefix
          content:
            application/json:
              schema:
                type: object
                required: [secret, prefix]
                properties:
                  secret:
                    type: string
                    example: "pidsk_9f2c…"
                  prefix:
                    type: string
                    example: "pidsk_9f2c41"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /apps/{id}/webhook:
    post:
      tags: [apps]
      summary: Set the app's ledger-callback endpoint
      description: |
        PeridotID POSTs signed ledger events here (e.g. `fiat.transfer.posted`
        when funds land on this app's escrow account). Returns a `pidwh_…` HMAC
        signing secret **once**; every callback carries
        `X-Pid-Signature: sha256=<hex>` over the raw body and `X-Pid-Event`.
        https required (http allowed for localhost). Owners only.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  example: "https://live2dev.com/peridot/webhook"
      responses:
        "200":
          description: The endpoint and its signing secret (shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                  secret:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [apps]
      summary: Remove the app's ledger-callback endpoint (and secret)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /apps/{id}/fees:
    get:
      tags: [apps]
      summary: Per-app fee schedule (topup / transaction / withdraw)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fee rows per operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /apps/{id}/fees/{operation}:
    put:
      tags: [apps]
      summary: Set one operation's per-app fee (stacks on the global fee)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: operation
          in: path
          required: true
          schema:
            type: string
            enum: [topup, transaction, withdraw]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [percentBps, minIdr, maxIdr, enabled]
              properties:
                percentBps:
                  type: integer
                minIdr:
                  type: string
                maxIdr:
                  type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: Saved fee
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /apps/{id}:
    patch:
      tags: [apps]
      summary: Update one of my apps
      description: Rename, change allowed origins, or disable the app. Owners only.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 60
                allowedOrigins:
                  type: array
                  maxItems: 20
                  description: Bare http(s) origins (no paths) — replaces the whole list
                  items:
                    type: string
                isActive:
                  type: boolean
      responses:
        "200":
          description: The updated app
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PidApp"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pid_access
      description: "JWT access token set by /auth/google/callback or /auth/refresh"
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 400
            message: ["handle must be 3-20 chars: lowercase letters, numbers, underscore"]
...
    Unauthorized:
      description: Missing, expired, or invalid access/refresh token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 401
            message: "Unauthorized"
...
    NotFound:
      description: The identity or profile does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            statusCode: 404
            message: "No Identity found"
...
  schemas:
...
    PidApp:
      type: object
      required: [id, clientId, ownerId, name, allowedOrigins, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          description: Public app id to pass as clientId in login/exchange calls
          example: "pidapp_0123456789abcdef0123456789abcdef"
        ownerId:
          type: string
        name:
          type: string
        allowedOrigins:
          type: array
          description: Managed list of bare origins — login redirects and CORS
          items:
            type: string
        isActive:
          type: boolean
        clientSecretPrefix:
          type: string
          nullable: true
          description: Identifying prefix of the current backend secret (the secret itself is never returned)
        clientSecretCreatedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string